V1.1

+++ iso/installer/install.sh
@@ -126,7 +126,7 @@ fi
 
 # Let's check if all dependencies are met
 function fuGET_DEPS {
-local myPACKAGES="apache2-utils apparmor apt-transport-https aufs-tools bash-completion build-essential ca-certificates cgroupfs-mount cockpit cockpit-docker curl dialog dnsutils docker.io docker-compose dstat ethtool fail2ban genisoimage git glances grc html2text htop ifupdown iptables iw jq libcrack2 libltdl7 lm-sensors man multitail net-tools npm ntp openssh-server openssl pass prips syslinux psmisc pv python-pip unattended-upgrades unzip vim wireless-tools wpasupplicant"
+local myPACKAGES="apache2-utils apparmor apt-transport-https aufs-tools bash-completion build-essential ca-certificates cgroupfs-mount cockpit cockpit-docker curl debconf-utils  dialog dnsutils docker.io docker-compose dstat ethtool fail2ban genisoimage git glances grc html2text htop ifupdown iptables iw jq libcrack2 libltdl7 lm-sensors man mosh multitail net-tools npm ntp openssh-server openssl pass prips software-properties-common syslinux psmisc pv python-pip unattended-upgrades unzip vim wireless-tools wpasupplicant"
 echo
 echo "### Bak /etc/apt/source.list To /etc/apt/source.list.bak"
 mv /etc/apt/sources.list /etc/apt/sources.list.bak
@@ -147,7 +147,10 @@ apt-get -y update
 echo
 echo "### Upgrading packages."
 echo
-apt-get -y dist-upgrade
+#Download and upgrade packages, but silently keep existing configs
+echo "docker.io docker.io/restart       boolean true" | debconf-set-selections -v
+echo "debconf debconf/frontend select noninteractive" | debconf-set-selections -v
+apt-get -y dist-upgrade -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confold" --force-yes
 echo
 echo "### Installing MoonStack dependencies."
 echo
@@ -338,7 +341,7 @@ if [ "$myMOON_DEPLOYMENT_TYPE" == "iso" ] || [ "$myMOON_DEPLOYMENT_TYPE" == "use
     "SENSOR" "Just Honeypots, EWS Poster & NSM" \
     "INDUSTRIAL" "Conpot, RDPY, Vnclowpot, ELK, NSM & Tools" \
     "COLLECTOR" "Heralding, ELK, NSM & Tools" \
-    "EXPERIMENTAL" "Experimental (Glutton instead of Honeytrap)" \
+    "NEXTGEN" "NextGen (Glutton instead of Honeytrap)" \
     "LEGACY" "Standard Edition from previous release" 3>&1 1>&2 2>&3 3>&-)
 fi
 
@@ -565,7 +568,7 @@ npm install https://gitee.com/stackw0rm/elasticsearch-dump.git -g 2>&1 | dialog 
 pip install --upgrade pip 2>&1 | dialog --title "[ Installing pip ]" $myPROGRESSBOXCONF
 hash -r 2>&1 | dialog --title "[ Installing pip ]" $myPROGRESSBOXCONF
 pip install elasticsearch-curator yq 2>&1 | dialog --title "[ Installing elasticsearch-curator, yq ]" $myPROGRESSBOXCONF
-git clone https://gitee.com/stackw0rm/moon.git /opt/moon 2>&1 | dialog --title "[ Cloning MoonStack ]" $myPROGRESSBOXCONF
+git clone --branch v1.1 --depth=1 https://gitee.com/stackw0rm/moon.git /opt/moon 2>&1 | dialog --title "[ Cloning MoonStack ]" $myPROGRESSBOXCONF
 cp /opt/moon/iso/installer/ctop-0.7.1-linux-amd64 /usr/bin/ctop 2>&1 | dialog --title "[ Installing ctop ]" $myPROGRESSBOXCONF
 chmod +x /usr/bin/ctop 2>&1 | dialog --title "[ Installing ctop ]" $myPROGRESSBOXCONF
 /opt/moon/iso/installer/set_mirrors.sh https://0at6ledb.mirror.aliyuncs.com 2>&1 | dialog --title "[ Set Docker Mirrors From Aliyun ]" $myPROGRESSBOXCONF
@@ -589,7 +592,8 @@ fi
 
 # Let's patch cockpit.socket, sshd_config
 sed -i 's#ListenStream=9090#ListenStream=64294#' /lib/systemd/system/cockpit.socket 2>&1 | dialog --title "[ Cockpit listen on tcp/64294 ]" $myPROGRESSBOXCONF
-sed -i 's#\#Port 22#Port 64295#' /etc/ssh/sshd_config 2>&1 | dialog --title "[ SSH listen on tcp/64295 ]" $myPROGRESSBOXCONF
+sed -i '/^port/Id' /etc/ssh/sshd_config && echo "Port 64295" >> /etc/ssh/sshd_config 2>&1 | dialog --title "[ SSH listen on tcp/64295 ]" $myPROGRESSBOXCONF
+echo "Port 64295" >> /etc/ssh/sshd_config 2>&1 | dialog --title "[ SSH listen on tcp/64295 ]" $myPROGRESSBOXCONF
 
 # Let's make sure only myCONF_MOON_FLAVOR images will be downloaded and started
 case $myCONF_MOON_FLAVOR in
@@ -609,9 +613,9 @@ case $myCONF_MOON_FLAVOR in
     echo "### Preparing COLLECTOR flavor installation."
     ln -s /opt/moon/etc/compose/collector.yml $myMOONCOMPOSE 2>&1>/dev/null
   ;;
-  EXPERIMENTAL)
-    echo "### Preparing EXPERIMENTAL flavor installation."
-    ln -s /opt/moon/etc/compose/experimental.yml $myMOONCOMPOSE 2>&1>/dev/null
+  NEXTGEN)
+    echo "### Preparing NEXTGEN flavor installation."
+    ln -s /opt/moon/etc/compose/nextgen.yml $myMOONCOMPOSE 2>&1>/dev/null
   ;;
   LEGACY)
     echo "### Preparing LEGACY flavor installation."
@@ -695,7 +699,7 @@ myCRONJOBS="
 */1 * * * *     root    mv --backup=numbered /data/dionaea/roots/ftp/* /data/dionaea/binaries/
 
 # Daily reboot
-27 3 * * *      root    systemctl stop moon && docker stop $(docker ps -aq) && docker rm $(docker ps -aq) && reboot
+27 3 * * *      root    systemctl stop moon && docker stop \$(docker ps -aq) || docker rm \$(docker ps -aq) || reboot
 
 # Check for updated packages every sunday, upgrade and reboot
 27 16 * * 0     root    apt-get autoclean -y && apt-get autoremove -y && apt-get update -y && apt-get upgrade -y && sleep 10 && reboot
@@ -755,10 +759,13 @@ tee -a /root/.bashrc 2>&1>/dev/null <<EOF
 $myROOTPROMPT
 PATH="$PATH:/opt/moon/bin"
 EOF
-tee -a /home/msec/.bashrc 2>&1>/dev/null <<EOF
+for i in $(ls -d /home/*/)
+   do
+ tee -a $i.bashrc 2>&1>/dev/null <<EOF
 $myUSERPROMPT
 PATH="$PATH:/opt/moon/bin"
 EOF
+done
 
 # Let's create ews.ip before reboot and prevent race condition for first start
 /opt/moon/bin/updateip.sh 2>&1>/dev/null
@@ -770,5 +777,10 @@ apt-get autoremove -y 2>&1 | dialog --title "[ Cleaning up ]" $myPROGRESSBOXCONF
 # Final steps
 cp /opt/moon/host/etc/rc.local /etc/rc.local 2>&1>/dev/null && \
 rm -rf /root/installer 2>&1>/dev/null && \
-dialog --no-ok --no-cancel --backtitle "$myBACKTITLE" --title "[ Thanks for your patience. Now rebooting. ]" --pause "" 6 80 2 && \
-reboot
+if [ "$myMOON_DEPLOYMENT_TYPE" == "auto" ];
+   then
+     echo "Done. Please reboot."
+   else
+     dialog --no-ok --no-cancel --backtitle "$myBACKTITLE" --title "[ Thanks for your patience. Now rebooting. ]" --pause "" 6 80 2 && \
+     reboot
+ fi
